refactor: Search 컴포넌트 폴더 구조 변경
