+++ src/components/common/Search/SearchFormWithList/index.tsx
@@ -1,15 +1,19 @@
 'use client';
-import SearchForm from '@/components/common/SearchForm/SearchForm';
-import SearchList from '../../common/SearchList/SearchList';
+import SearchForm from '@/components/common/Search/SearchForm';
+import SearchList from '../SearchList';
 import { SearchMonster } from '@/model/monster';
-import styles from './css/Search.module.css';
-import useSearch from '@/hooks/useSearch';
+import styles from './css/index.module.css';
+import useSearch from '@/components/common/Search/hooks/useSearch';
 
 type Props = {
   monsters: SearchMonster[];
+  smHidden?: 'sm-hidden' | false;
 };
 
-export default function Search({ monsters }: Props) {
+export default function SearchFormWithList({
+  monsters,
+  smHidden = 'sm-hidden',
+}: Props) {
   const {
     text,
     searchRef,
@@ -26,7 +30,10 @@ export default function Search({ monsters }: Props) {
   } = useSearch(monsters);
 
   return (
-    <div className={`sm-hidden ${styles.search}`} ref={searchRef}>
+    <div
+      className={`${smHidden ? smHidden : undefined} ${styles.search}`}
+      ref={searchRef}
+    >
       <SearchForm
         text={text}
         select={select}
